requirements.txt
asgiref==3.7.1
attrs==23.1.0
Django==4.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
drf-spectacular==0.26.2
future==0.18.3
inflection==0.5.1
jsonschema==4.17.3
psycopg2==2.9.6
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
uritemplate==4.1.1
